% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tables.R
\name{calculate_tables}
\alias{calculate_tables}
\title{Create a list tables of OD and Concentration changes through time from the log files.}
\usage{
calculate_tables(
  dil_check = T,
  bc_check = T,
  C0 = 0,
  CS1 = 0,
  CS2 = 10,
  V = 20,
  dV = 4,
  od_file = "6-21-17-OD.txt",
  pump_file = "6-21-17-pumps.txt",
  dil_file = "dilutions.txt",
  bc_file = "bottle_change.txt",
  tubes_selected = 1:6,
  zero = c(Tube1 = 0, Tube2 = 0, Tube3 = 0, Tube4 = 0, Tube5 = 0, Tube6 = 0),
  dV.change = c()
)
}
\arguments{
\item{dil_check}{- Were any manual dilutions introdused? Default is TRUE. If TRUE - need to provide dil_file}

\item{bc_check}{- Was concentration of a drug changed during the run. Default is TRUE.}

\item{C0}{- Concentration in reactors at the start of the run.}

\item{CS1}{- Concentration in Pump 1 bottle at the start of the run.}

\item{CS2}{- Concentration in Pump 2 bottle at the start of the run.}

\item{V}{- Volume of of media in the reactors.}

\item{dV}{- The delution volume.}

\item{od_file}{- Log file of OD measurements.}

\item{pump_file}{- Log file of pumps activation.}

\item{dil_file}{- File that specifies dilution time, ratio (volume taken to total volume), concentration of the drug in the added media and the reactor.}

\item{bc_file}{- File that specifies bottle change time, concentration in the Pump 1 bottle, Pump 2 bottle.}

\item{tubes_selected}{- Vector with numbers of reactors activated in the current run.}

\item{zero}{- zero levels of each tube. The levels should be in the increasing order.}

\item{dV.change}{- file with time of change and new dV}
}
\description{
Create a list tables of OD and Concentration changes through time from the log files.
}
