% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.to.middle.R
\name{collapse.to.middle}
\alias{collapse.to.middle}
\title{Make tables with measurement only in the middle of interval.}
\usage{
collapse.to.middle(tables)
}
\arguments{
\item{tables}{- List of tables that were created by \emph{calaculate_tables} function.}
}
\description{
Make tables with measurement only in the middle of interval.
}
