% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative}
\alias{cumulative}
\title{Cumulative values of OD, Concentration, number of generations and replication events.}
\usage{
cumulative(tables, KOD = 8 * 10^8, V = 20)
}
\arguments{
\item{tables}{- List of tables that were created by \emph{calaculate_tables} function.}

\item{KOD}{- coefficient of the OD to number of cell conversion.}

\item{V}{- volume of media in the reactors.}
}
\description{
Cumulative values of OD, Concentration, number of generations and replication events.
}
