% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw OD/time and Concentration/time curves.}
\usage{
draw(
  tables,
  c.units = "uM",
  MIC = 0.016,
  log2 = T,
  p.size = 0.5,
  add.model = F,
  model.table = data.frame(),
  m.prob = "",
  params = list(),
  dist = 2,
  ymeta = 0.3,
  sample_color = "red",
  bottle_color = "blue",
  repel = F,
  legend = T,
  m1 = T
)
}
\arguments{
\item{tables}{- List of tables that were created by \emph{calaculate_tables} function.}

\item{c.units}{- Text variable that specifies concentration units.}

\item{MIC}{- Minimum Inhibitory Concentration.}

\item{log2}{- Should the Concentration axes transformed with log2 transformation? Default value is TRUE (log2 coordinates).}

\item{p.size}{- Size of points on the OD plot.}

\item{add.model}{- If mutation dynamics modelling was made with the \emph{mutability} function TRUE value will add this model to OD plot. Default value is FALSE.}

\item{model.table}{- Table with the model dynamics produced by \emph{mutability} function. Could be omitted if \strong{add.model == FALSE}.}

\item{m.prob}{- Mutation probability. Could be omitted if \strong{add.model == FALSE}. Calculated inside the \emph{calaculate_tables} function.}

\item{params}{- information about dilutions and bottle changes geneated by \emph{get_draw_parameters}}

\item{dist}{- distance between lines indicated sample collections or bottle changes and text}

\item{ymeta}{- height of bottle change and sample collection lables}

\item{sample_color}{- color for lines indicating sampling (default: 'red')}

\item{bottle_color}{- color for lines indicating bottle change (default: 'blue')}

\item{repel}{- use repel labels instead of line labels (default: FALSE)}

\item{legend}{- show legend (deafult: TRUE)}

\item{m1}{- if TRUE plot concentration in media 1 (default: TRUE)}
}
\description{
Draw OD/time and Concentration/time curves.
}
