% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomSeries.R
\name{geomSeries}
\alias{geomSeries}
\title{Calculate a series of numbers in a geometric progression
with a given "base" up to "max" value.}
\usage{
geomSeries(base, max)
}
\arguments{
\item{base}{- A base of the geometric progression}

\item{max}{- A number that geometric progression should not exceed.}
}
\description{
Calculate a series of numbers in a geometric progression
with a given "base" up to "max" value.
}
