% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draw_parameters.R
\name{get_draw_parameters}
\alias{get_draw_parameters}
\title{Create a list of auxillary parameters for draw function}
\usage{
get_draw_parameters(dil_check = T, bc_check = T, CS1 = 0, CS2 = 10,
  od_file = "6-21-17-OD.txt", dil_file = "dilutions.txt",
  bc_file = "bottle_change.txt")
}
\arguments{
\item{dil_check}{- Were any manual dilutions introdused? Default is TRUE. If TRUE - need to provide dil_file}

\item{bc_check}{- Was concentration of a drug changed during the run. Default is TRUE.}

\item{CS1}{- Concentration in Pump 1 bottle at the start of the run.}

\item{CS2}{- Concentration in Pump 2 bottle at the start of the run.}

\item{od_file}{- Log file of OD measurements.}

\item{dil_file}{- File that specifies dilution time, ratio (volume taken to total volume), concentration of the drug in the added media and the reactor.}

\item{bc_file}{- File that specifies bottle change time, concentration in the Pump 1 bottle, Pump 2 bottle.}
}
\description{
Create a list of auxillary parameters for draw function
}
