% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.rates.R
\name{growth.rates}
\alias{growth.rates}
\title{Calculate list of table for growth rates (h^-1) at each interval with concentration on each interval, start and stop time for interval.}
\usage{
growth.rates(tables, V = 20, MIC = 0.016)
}
\arguments{
\item{tables}{- List of tables that were created by \emph{calaculate_tables} function.}

\item{V}{- Volume of of media in the reactors.}

\item{MIC}{- Minimum Inhibitory Concentration.}
}
\description{
Calculate list of table for growth rates (h^-1) at each interval with concentration on each interval, start and stop time for interval.
}
