% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.rep.events.R
\name{num.rep.events}
\alias{num.rep.events}
\title{Calculate number of events of replication}
\usage{
num.rep.events(N.start, N.final)
}
\arguments{
\item{N.start}{- Number of cells at the beginning of the interval.}

\item{N.final}{- Number of cells at the end of the interval.}
}
\description{
Calculate number of events of replication
}
