% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od.plot2log.R
\name{od.plot2log}
\alias{od.plot2log}
\title{Convert plot save format to log format.}
\usage{
od.plot2log(plotfile = "OD0.txt", logfile = "OD_log.txt")
}
\arguments{
\item{plotfile}{- file with plot save. default = OD_plot.txt}

\item{logfile}{- path where log format should be saved. default = OD_log.txt}
}
\description{
Convert plot save format to log format.
}
