% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.m.tables.R
\name{write.m.tables}
\alias{write.m.tables}
\title{Write tables to file. Each reactor will have it's own files.}
\usage{
write.m.tables(tables, united = F, col = F)
}
\arguments{
\item{tables}{- List of tables generated by \emph{calaculate_tables} function.}

\item{united}{- If TRUE creates the united table (*_United.txt) for OD and Concentration.}

\item{col}{- If TRUE the input table is collapsed
If FALSE - files for the tables of OD and Concetration changes will be created separately (*_OD.txt, *_Conc.txt)}
}
\description{
Write tables to file. Each reactor will have it's own files.
}
